Fixed Mathpad dragging bug and hid Mathpad when no activity selected
Moved the mousemove and mouseup handlers off of the div and onto the window so dragging doesn't get stuck if the mouse moves too fast.

Also hid Mathpad when no activity is selected because opening Mathpad logs an event which causes an error because the logger doesn't have the activityName variable set.