Observe xhr success; notify user if offline; retry xhr if failure
The adds two features: Notifying the user if they go offline, and
resending missed log events when the user comes back online.

When we send a POST message for a log event, we check its success and
if it fails we notify the user and then keep retrying every 2 seconds
until success.

We could also observe posts to Firebase, but every post to FB should also
include a log event, so it would probably be redundant. We don't bother to
explicitly re-send the FB events, as the FB library seems to do this
automatically.

Note that if the Log Manager goes down, users would be told they are offline.