Slide MathPad in the first time we open it.
Use a css transition to do sliding. Add a new state, 'unopened', which
determines whether to use the css transitions (or all dragging would
have the transition applied).

This also required delaying the initial input focus, or the browser
would scroll to keep the math pad in the view as it slid in.