Change calculator buttons and restyle
This makes the main panel a simple four-function calculator, with
memory buttons above along with two extra operators (sqrt and inv).
Backspace and plusMinus are removed to simplify the operators.

Styled to look a little closer to something like the Casio Personal M-1.

Memory functions and changes to key characters not added yet.