Added pin label component [#126223245]
This allows for dropping the probe on pin label instead of just the pin.  The wire component was also updated so that it doesn't consume mouse events while the probe is moving so that you can drop the probe onto a pin with one or more wires already attached to it.