Fixed chip deletes and chip counts [#140139707]
The remote chip delete path was not calling removeComponent() which it now is.   In some cases this caused the chip to be removed from the board but all the wires attached to the chip would remain.  Also moved the logic chip drawer chip counting code directly into the view as a computed value so that when a client reconnects the chip count is not reset to 0.