Find nearest valid hole when placing chip
When dragging in a new chip or moving an existing one, we find the
nearest hole to the top-left pin that would allow us to fit the chip
on the board, spanning the central gutter. We don't worry about users
who place a chip in an invalid position, we simply always make it valid.