Prevent using occupied holes in breadboard.
Add `.connected` to make use of the existing connected status for
wires, and don't allow dragging from or to connected holes.

For chips, every time we add a chip we set all the holes were are
plugged into, plus the holes in-between (covered by the body) as
connected. We need to also retain this info in the component, so we
can disconnect them when we move the component.

If we drag to an invalid place, we either throw away the component (for
a new drag) or snap it back to its original location.