Changed model to calculate powered circuits instead of "connected" circuits.
Previously the logic chips were assumed to have Vcc and ground wired and used a "connected" flag to denote when holes and pins were wired.  The model has now been changed to remove the connected flag and replace it with a powered flag that (when a breadboard is present) is only set if the circuit has either a source voltage or ground voltage in its network.  The PIC activity uses the powered flag also but it overrides the search for the source or ground voltage.