Change version number to build number [#120136869]
Instead of using the package.json version as the displayed version
number in the footer the first 8 characters of the commit hash is shown
as a link to the commit on GitHub.  This was previously a hidden link in
the footer.

The full commit hash is also recoreded under "build" in the log manager
parameters.